Cambios en Caejros y Identity (Autorizacion y bloqueo >> aun no funciona 100%)

+++ CasinoRegistro/Areas/Admin/Controllers/CajerosController.cs
@@ -15,34 +15,55 @@ using CasinoRegistro.DataAccess.Data.Initialiser;
 using CasinoRegistro.DataAccess.Helpers;
 using Microsoft.VisualStudio.Web.CodeGenerators.Mvc.Templates.BlazorIdentity.Pages.Manage;
 using Microsoft.AspNetCore.Identity;
+using CasinoRegistro.Utilities;
+using System.Security.Claims;
+using Microsoft.AspNetCore.Authorization;
+using Azure.Identity;
 
 namespace CasinoRegistro.Areas.Admin.Controllers
 {
+    [Authorize(Roles = "Administrador,Secretaria")]
     [Area("Admin")]
     public class CajerosController : Controller
-    {       
+    {
 
         private readonly IContenedorTrabajo _contenedorTrabajo;
         private readonly IWebHostEnvironment _hostingEnvironment; // para acceder a las carpetas del proyecto, para guardar/editar la imagen      
         private readonly UserManager<IdentityUser> _userManager;
         private readonly CasinoRegistroDbContext _db;
+        private readonly RoleManager<IdentityRole> _roleManager;
 
-        public CajerosController(IContenedorTrabajo contenedorTrabajo, IWebHostEnvironment hostingEnvironment,          
-            UserManager<IdentityUser> userManager, CasinoRegistroDbContext db)
+
+        public CajerosController(IContenedorTrabajo contenedorTrabajo, IWebHostEnvironment hostingEnvironment,
+            UserManager<IdentityUser> userManager, CasinoRegistroDbContext db, RoleManager<IdentityRole> roleManager)
         {
             _contenedorTrabajo = contenedorTrabajo;
-            _hostingEnvironment = hostingEnvironment;           
+            _hostingEnvironment = hostingEnvironment;
             _userManager = userManager;
             _db = db;
+            _roleManager = roleManager;
         }
 
         // GET: Admin/Cajeros
         public async Task<IActionResult> Index()
         {
-          
-           return View();
-        }
 
+            //Opción 1: Obtener todos los usuario
+            return View(_contenedorTrabajo.Cajero.GetAll());
+
+            //Opción 2: Obtener todos los usuarios menos el que esté logueado, para no bloquearse el mismo
+
+
+            //var claimsIdentity = (ClaimsIdentity)this.User.Identity;
+            //var usuarioActual = claimsIdentity.FindFirst(ClaimTypes.NameIdentifier);
+
+
+            // var user = await _userManager.FindByNameAsync(claimsIdentity.Name);
+
+            //// return View(_contenedorTrabajo.Cajero.GetAll());
+
+            //return View(_contenedorTrabajo.Cajero.GetAll(u => u.Id != usuarioActual.Value));
+        }
 
 
         // GET: Admin/Cajeros/Details/5
@@ -55,7 +76,7 @@ namespace CasinoRegistro.Areas.Admin.Controllers
             };
 
             if (id != null)
-            {            
+            {
                 cajeroViewModel.CajeroUserVM = _contenedorTrabajo.Cajero.Get(id.GetValueOrDefault());
             }
 
@@ -66,11 +87,11 @@ namespace CasinoRegistro.Areas.Admin.Controllers
         // GET: Admin/Cajeros/Create
         [HttpGet]
         public IActionResult Create()
-        {  
+        {
             CajeroViewModel cajeroViewModel = new CajeroViewModel()
-            { 
+            {
                 CajeroUserVM = new CasinoRegistro.Models.CajeroUser(),
-               // ListaPlataformas = _contenedorTrabajo.Plataforma.GetListaPlataformas()
+                // ListaPlataformas = _contenedorTrabajo.Plataforma.GetListaPlataformas()
             };
             return View(cajeroViewModel);
         }
@@ -80,8 +101,8 @@ namespace CasinoRegistro.Areas.Admin.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create(CajeroViewModel cajeroVM)        
-        {          
+        public async Task<IActionResult> Create(CajeroViewModel cajeroVM)
+        {
 
             if (ModelState.IsValid)
             {
@@ -91,42 +112,80 @@ namespace CasinoRegistro.Areas.Admin.Controllers
                     try
                     {
                         string rutaPrincipal = _hostingEnvironment.WebRootPath;
-                     var archivo = HttpContext.Request.Form.Files;
+                        var archivo = HttpContext.Request.Form.Files;
 
-                    if (cajeroVM.CajeroUserVM.Id == 0 && archivo.Count() > 0)
-                    {
-                        //Nuevo Cajero
-                        string nombreArchivo = Guid.NewGuid().ToString();   //como es nuevo el cajero, le pongo un Guid como nombre
-                        var subidas = Path.Combine(rutaPrincipal, @"imagenes\cajeros"); //accederá a la carpeta en wwwroot
-                        var extension = Path.GetExtension(archivo[0].FileName);
-
-                        using (var fileStreams = new FileStream(Path.Combine(subidas, nombreArchivo + extension), FileMode.Create))
+                        if (cajeroVM.CajeroUserVM.Id == 0 && archivo.Count() > 0)
                         {
-                            archivo[0].CopyTo(fileStreams);
+                            //Nuevo Cajero
+                            string nombreArchivo = Guid.NewGuid().ToString();   //como es nuevo el cajero, le pongo un Guid como nombre
+                            var subidas = Path.Combine(rutaPrincipal, @"imagenes\cajeros"); //accederá a la carpeta en wwwroot
+                            var extension = Path.GetExtension(archivo[0].FileName);
+
+                            using (var fileStreams = new FileStream(Path.Combine(subidas, nombreArchivo + extension), FileMode.Create))
+                            {
+                                archivo[0].CopyTo(fileStreams);
+                            }
+
+                            //guardo la ruta en la base de datos
+                            cajeroVM.CajeroUserVM.UrlImagen = @"imagenes\cajeros\" + nombreArchivo + extension;
+
                         }
 
-                        //guardo la ruta en la base de datos
-                        cajeroVM.CajeroUserVM.UrlImagen = @"imagenes\cajeros\" + nombreArchivo + extension;
+                        ////Logica para guardar en BD
+                        //_contenedorTrabajo.Cajero.Add(cajeroVM.CajeroUserVM);
 
-                    }
 
-                    //Logica para guardar en BD
-                    _contenedorTrabajo.Cajero.Add(cajeroVM.CajeroUserVM);
+                        //_contenedorTrabajo.Save();
 
-                    
-                        _contenedorTrabajo.Save();                      
 
                         //CREA EL USUARIO EN ASPNET USERS
-                        _userManager.CreateAsync(new IdentityUser
+
+                        IdentityUser user = new IdentityUser();
+
+                        user.UserName = cajeroVM.CajeroUserVM.Email;
+                        user.Email = cajeroVM.CajeroUserVM.Email;
+                        user.EmailConfirmed = true;
+                        _userManager.CreateAsync(user, "Admin1234.").GetAwaiter().GetResult();
+
+
+                        //Aquí validamos si los roles existen sino se crean
+                        if (!await _roleManager.RoleExistsAsync(CNT.Administrador))
                         {
-                            UserName = cajeroVM.CajeroUserVM.Email,
-                            Email = cajeroVM.CajeroUserVM.Email,
+                            await _roleManager.CreateAsync(new IdentityRole(CNT.Administrador));
+                            await _roleManager.CreateAsync(new IdentityRole(CNT.Secretaria));
+                            await _roleManager.CreateAsync(new IdentityRole(CNT.Cajero));
+                        }
 
-                            EmailConfirmed = true,
-                        }, "Admin1234.").GetAwaiter().GetResult();
+                        //Obtenemos el rol seleccionado
+                        string rol = Request.Form["radUsuarioRole"].ToString();
 
-                        // CajeroUser usuario = _db.Cajero.Where(us => us.Email == cajeroVM.CajeroUserVM.Email).FirstOrDefault();
+                        //Validamos si el rol seleccionado es Admin y si lo es lo agregamos
+                        if (rol == CNT.Administrador)
+                        {
+                            await _userManager.AddToRoleAsync(user, CNT.Administrador);
 
+                        }
+                        else
+                        {
+                            if (rol == CNT.Secretaria)
+                            {
+                                await _userManager.AddToRoleAsync(user, CNT.Secretaria);
+                            }
+                            else
+                            {
+                                await _userManager.AddToRoleAsync(user, CNT.Cajero);
+                            }
+
+                        }
+
+                        cajeroVM.CajeroUserVM.Rol = rol;
+                        cajeroVM.CajeroUserVM.Estado = true;
+
+                        //Logica para guardar en BD
+                        _contenedorTrabajo.Cajero.Add(cajeroVM.CajeroUserVM);
+
+
+                        _contenedorTrabajo.Save();
 
                         transaction.Commit();
                         return RedirectToAction(nameof(Index));
@@ -155,12 +214,13 @@ namespace CasinoRegistro.Areas.Admin.Controllers
                             }
                         }
 
+                        return View(cajeroVM);
                     }
                 }
             }
 
             //para no perder los datos de la lista de plataformas
-           //cajeroVM.ListaPlataformas = _contenedorTrabajo.Plataforma.GetListaPlataformas();
+            //cajeroVM.ListaPlataformas = _contenedorTrabajo.Plataforma.GetListaPlataformas();
 
             return View(cajeroVM);
         }
@@ -196,6 +256,8 @@ namespace CasinoRegistro.Areas.Admin.Controllers
                 {
                     try
                     {
+
+
                         string rutaPrincipal = _hostingEnvironment.WebRootPath;
                         var archivos = HttpContext.Request.Form.Files;
 
@@ -209,12 +271,18 @@ namespace CasinoRegistro.Areas.Admin.Controllers
                             var extension = Path.GetExtension(archivos[0].FileName);
                             var nuevaExtension = Path.GetExtension(archivos[0].FileName);
 
-                            var rutaImagen = Path.Combine(rutaPrincipal, cajeroDesdeBd.UrlImagen.TrimStart('\\'));
 
-                            if (System.IO.File.Exists(rutaImagen))
+                            if (cajeroDesdeBd.UrlImagen != null)
                             {
-                                System.IO.File.Delete(rutaImagen);
+
+                                var rutaImagen = Path.Combine(rutaPrincipal, cajeroDesdeBd.UrlImagen.TrimStart('\\'));
+
+                                if (System.IO.File.Exists(rutaImagen))
+                                {
+                                    System.IO.File.Delete(rutaImagen);
+                                }
                             }
+
 
                             //Nuevamente subimos el archivo
                             using (var fileStreams = new FileStream(Path.Combine(subidas, nombreArchivo + extension), FileMode.Create))
@@ -224,9 +292,9 @@ namespace CasinoRegistro.Areas.Admin.Controllers
 
                             cajeroVM.CajeroUserVM.UrlImagen = @"\imagenes\cajeros\" + nombreArchivo + extension;
 
-                          
 
-                            return RedirectToAction(nameof(Index));
+
+                           
                         }
                         else
                         {
@@ -234,22 +302,65 @@ namespace CasinoRegistro.Areas.Admin.Controllers
                             cajeroVM.CajeroUserVM.UrlImagen = cajeroDesdeBd.UrlImagen;
                         }
 
+                        
+
+                        var user = await _userManager.FindByNameAsync(cajeroDesdeBd.Email);
+
                         //EDITO EL USUARIO EN ASPNETUSER
-                        if (cajeroDesdeBd.Email != cajeroVM.CajeroUserVM.Email)
+                        if (user != null)
                         {
-                           
-                            var user = await _userManager.FindByEmailAsync(cajeroDesdeBd.Email);
-                            if (user != null)
+                            if (cajeroDesdeBd.Email != cajeroVM.CajeroUserVM.Email)
                             {
+                           
                                 var setEmailResult = await _userManager.SetEmailAsync(user, cajeroVM.CajeroUserVM.Email);
                                 var setUserNameResult = await _userManager.SetUserNameAsync(user, cajeroVM.CajeroUserVM.Email);
+
+
                                 if (setEmailResult.Succeeded && setUserNameResult.Succeeded)
                                 {
                                     await _userManager.UpdateAsync(user);
                                 }
                             }
 
+                            //Obtenemos el rol seleccionado
+                            string rol = Request.Form["radUsuarioRole"].ToString();
+
+                            //EDITO EL ROL SI CORRESPONDE
+                            if (cajeroDesdeBd.Rol != rol)
+                            {
+                                var roles = await _userManager.GetRolesAsync(user);
+                                var result = await _userManager.RemoveFromRoleAsync(user, roles[0]);
+
+                                //Validamos si el rol seleccionado es Admin y si lo es lo agregamos
+                                if (rol == CNT.Administrador)
+                                {
+                                    await _userManager.AddToRoleAsync(user, CNT.Administrador);
+                                }
+                                else
+                                {
+                                    if (rol == CNT.Secretaria)
+                                    {
+                                        await _userManager.AddToRoleAsync(user, CNT.Secretaria);
+                                    }
+                                    else
+                                    {
+                                        await _userManager.AddToRoleAsync(user, CNT.Cajero);
+                                        //TODO: CORREGIR CUANDO PUEDA HACER QUE EL EDIT MUESTRE LO ELEGIDO TAMBIEN
+                                        cajeroVM.CajeroUserVM.Rol = "Cajero";
+                                    }
+                                }
+                                cajeroVM.CajeroUserVM.Rol = rol;
+                            }
+                            else
+                            {
+                                //TODO: CORREGIR CUANDO PUEDA HACER QUE EL EDIT MUESTRE LO ELEGIDO
+                                await _userManager.AddToRoleAsync(user, CNT.Cajero);
+                                cajeroVM.CajeroUserVM.Rol = "Cajero";
+                            }
+
+
                         }
+
 
                         _contenedorTrabajo.Cajero.Update(cajeroVM.CajeroUserVM);
                         _contenedorTrabajo.Save();
@@ -281,6 +392,9 @@ namespace CasinoRegistro.Areas.Admin.Controllers
                                 ModelState.AddModelError(string.Empty, "Contacte con el administrador >> Error: " + ex.InnerException.Message);
                             }
                         }
+                     
+
+                        return View(cajeroVM);
                     }
                 }
 
@@ -338,18 +452,70 @@ namespace CasinoRegistro.Areas.Admin.Controllers
         }
 
 
-        [HttpDelete]
-        public IActionResult Delete(int id)
+        //[HttpDelete]
+        //public IActionResult Delete(int id)
+        //{
+        //    var objFromDb = _contenedorTrabajo.Cajero.Get(id);
+        //    if (objFromDb == null)
+        //    {
+        //        return Json(new { success = false, message = "Error borrando el cajero" });
+        //    }
+
+        //    _contenedorTrabajo.Cajero.Remove(objFromDb);
+        //    _contenedorTrabajo.Save();
+        //    return Json(new { success = true, message = "Cajero Borrado Correctamente" });
+        //}
+
+        public async Task<IActionResult> BloquearDesloquearCajero(int id)
         {
-            var objFromDb = _contenedorTrabajo.Cajero.Get(id);
+            var cajeroDesdeBd = _contenedorTrabajo.Cajero.Get(id);
+
+            var objFromDb = _userManager.FindByNameAsync(cajeroDesdeBd.Email);
+
+         
+          
+
+            //var objFromDb = _contenedorTrabajo.Cajero.GetString(id);
             if (objFromDb == null)
             {
-                return Json(new { success = false, message = "Error borrando el cajero" });
+                return Json(new { success = false, message = "Error bloqueando el cajero." });
             }
 
-            _contenedorTrabajo.Cajero.Remove(objFromDb);
-            _contenedorTrabajo.Save();
-            return Json(new { success = true, message = "Cajero Borrado Correctamente" });
+            //en un solo metodo, bloqueo o desbloqueo el usuario segun corresponda
+
+            if (objFromDb.Result.LockoutEnd == null || objFromDb.Result.LockoutEnd < DateTime.Now)
+            {
+
+                //_contenedorTrabajo.Cajero.BloquearCajero(cajeroDesdeBd,objFromDb);
+
+               // objFromDb.Result.LockoutEnd = DateTime.Now.AddYears(1000);
+
+                cajeroDesdeBd.Estado = false;
+                
+               await _userManager.SetLockoutEndDateAsync(objFromDb.Result, DateTime.Now.AddYears(1000));
+               await _userManager.UpdateAsync(objFromDb.Result);
+                _contenedorTrabajo.Save();
+               
+                return Json(new { success = true, message = "Cajero Bloqueado Correctamente" });
+            }
+
+            else
+            {
+
+                // _contenedorTrabajo.Cajero.DesloquearCajero(cajeroDesdeBd,objFromDb);
+
+                //cajeroDesdeBd.Estado = false;
+
+                await _userManager.SetLockoutEndDateAsync(objFromDb.Result, null);
+                await _userManager.UpdateAsync(objFromDb.Result);
+                cajeroDesdeBd.Estado = true;
+
+                _contenedorTrabajo.Save();
+                return Json(new { success = true, message = "Cajero DesBloqueado Correctamente" });
+            }
+            // return View(user);
+
+
         }
 
         #endregion
